Add a README for the data files
This README explains the data files - what information
they contain, what variables are in them, what those variables
mean, and which figures the files are relevant to.

The purpose of the README is to help make it easier for people
to understand the files, thereby allowing them to use the
data. (For instance, in reproducing the figures.)