Fix mistake in data file
For the models file, I had been computing
the result for Robin and I's result in a dumb
fashion - if the total from the kite as NaN,
I simply put the total as that which came
from the "L", instead of making it NaN as well.
This made the plots look a bit stupid. I fixed
this by basically regenerating the data file,
and making sure that any values which should
be NaN in our model, are.