Add more text in Section V
This commit introduces a somewhat substantial
content change in Section V. This section
is suppose to compare our theory to numerical
experiments; at present, that comparison simply
involves making a plot of <LLRS> versus d and
rank(rho0).

However, our theory *also* has another component-
the average decrement (z). A reader could look at
the comparison as-is, and ask themselves "Well, **why**
does your result start to break down as r -> d?".

I attempt to answer that question by also providing numerical
results for z itself. It turns out our expression for z is actually
quite inaccurate as r->d. (By which I mean the value we predict,
and the value we obtain by numerical averaging, are dramatically
different.)

However, if you simply take the numerical average for z, and plug
it into our formula, you get a prediction which is extremely accurate!
This implies that the main reason why our prediction for <LLRS> is off
**is not** because our formula is incorrect; rather, it's because
our expression for z is off! This is a point worth making, as it
clears us of any charges that we got <LLRS> wrong. What's more,
it also opens up the door to suggesting improvements could be made,
or (in a later paper) cautioning people against blithely taking
that formula and using it in their context. (This would be useful
when, for example, we write something about approximating <Rank(rhoML)>.)