Re-work proof in Section III B for solid tangent cone
The previous approach defined the tangent cone --
which is a surface -- and then was left with the issue of
how to "fill it in" with respect to the local state space.

This commit changes the approach, and instead takes the limit
of expanding entire convex sets around rho_0 (defined as the
intersection of the model and the ball), thereby circumventing
that problem.