Update solution for z (typical decrement)
I've been working this week to re-derive RBK's result
for z, and was successful in doing so. In turn, I found
there's a higher-order term that, if you include it
in the expansion of z for N->oo, you can make z behave
better when r -> d. This commit updates Equation (23)
to include that extra term, plus refactors the expressions
for the coefficients.