Change N->n in Section IV
One of the referee reports suggested
we change N in IV, because we use
it earlier in the paper. Because
N was quantifies the dimension of the
"null space" of rho_0, we chose to
use n.

These changes also propagate to the
second supplemental NB, where we derive
the approximate solution for z.